Use Floating Point instead of Integers

+++ lib/SensorFusion/src/IKalmanFilter.h
@@ -60,9 +60,9 @@ public:
     /** Results of the Sensor Fusion Algorithm. */
     typedef struct _PositionData
     {
-        int32_t positionX; /**< Current Position in x direction in mm. */
-        int32_t positionY; /**< Current Position in y direction in mm. */
-        int32_t angle;     /**< Current Orientation in in mrad. */
+        float positionX; /**< Current Position in x direction in mm. */
+        float positionY; /**< Current Position in y direction in mm. */
+        float angle;     /**< Current Orientation in in mrad. */
     } PositionData;
 
     /**
